requirements.txt
djangorestframework==3.16.0
Django==5.2.3
pillow==11.3.0
django-jazzmin==3.0.1